+++ vega-dataflow/src/main/kotlin/dev/aster/vega/dataflow/transform/CrossFilter.kt
@@ -1,7 +1,10 @@
+@file:OptIn(InternalAsterVegaApi::class)
+
 package dev.aster.vega.dataflow.transform
 
 import dev.aster.vega.expression.JsSemantics
 import dev.aster.vega.model.DiagnosticCodes
+import dev.aster.vega.model.InternalAsterVegaApi
 import dev.aster.vega.model.VegaValue
 import dev.aster.vega.model.field